+++ qiskit/providers/ibmq/api/clients/account.py
@@ -29,6 +29,7 @@ from ..exceptions import (RequestsApiError, WebsocketError,
                           WebsocketTimeoutError, UserTimeoutExceededError)
 from ..rest import Api
 from ..session import RetrySession
+from ..exceptions import ApiIBMQProtocolError
 from .base import BaseClient
 from .websocket import WebsocketClient
 
@@ -288,11 +289,17 @@ class AccountClient(BaseClient):
 
         Returns:
             job information.
+
+        Raises:
+            ApiIBMQProtocolError: if an unexpected result is received from the server.
         """
         if use_object_storage:
             return self._job_result_object_storage(job_id)
 
-        return self.job_get(job_id)['qObjectResult']
+        try:
+            return self.job_get(job_id)['qObjectResult']
+        except KeyError as err:
+            raise ApiIBMQProtocolError(str(err))
 
     def _job_result_object_storage(self, job_id: str) -> Dict:
         """Retrieve and return a result using object storage.